Reducing code duplication to minimum.
